requirements-dev.txt
# page-generation
gh-issues

# linting
pre-commit
typer

# Automation
langchain
langchain_community
langchain_ollama
rich
python-frontmatter

# Testing
ephemeral_port_reserve
pytest-playwright
pytest-xprocess
